pyproject.toml
[tool.poetry]
name = "quranref"
version = "2.0.0"
description = ""
authors = ["Kashif Iftikhar <kashif@compulife.com.pk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.2"}
pydantic-settings = "^2.4.0"
arango-orm = "^1.0.2"
typer = "^0.12.5"
jinja2 = "^3.1.4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"
bandit = "^1.7.9"
ipython = "^8.27.0"
pre-commit = "^3.8.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pytest-env = "^1.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
quranref-cli = 'quranref.cli:main'

[tool.bandit]
exclude_dirs = ["tests"]
# B101 = ignore assert_used
# B404 = allow use of sub process module
# B603 = allow calls to Popen with shell=False
skips = ["B101", "B404", "B603"]

[tool.ruff]
line-length = 100
src = ["quranref"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--cov=quranref --cov-fail-under=50"
testpaths = [
    "tests",
]

filterwarnings = [
    "ignore::DeprecationWarning",
]

env = [
    "QURANREF_TESTING=true"
]
